feat(18): basic dmn structs and simple evaluation refactor 2
