+++ pkg/dmn/dmn_util.go
@@ -1,8 +1,6 @@
 package dmn
 
-import (
-	"github.com/pbinitiative/zenbpm/pkg/storage/dmn"
-)
+import "github.com/pbinitiative/zenbpm/pkg/dmn/model/dmn"
 
 func findDecision(dmnDefinition *DecisionDefinition, decisionId string) *dmn.TDecision {
 	for _, decision := range dmnDefinition.definitions.Decisions {
